task_2/requirements.txt
asyncpg==0.27.0
alembic==1.11.0
fastapi==0.95.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.30.0
SQLAlchemy==2.0.13
uvicorn==0.22.0
